acceot `impl IntoCStr` for all public APIs taking strings
...and remove the "&str happens to be null terminated" hack.

Now that it's easy to just write `c"a c str literal"`, we don't need
this hack anymore, and we can just accept anything that can be converted
into a CStr.

This generally isn't a pessimization even for `format!("...\0")`,
since format! generally ends up with some extra space in the buffer, so
we can append the null character basically for free. (And of course it
makes client code much less janky, since we can get rid of all the
hardcoded `\0`.)

This isn't done for emitting events yet, since that will take a bit more
work to accept e.g. tuples with mixed types, something that will be
pretty common.